начал ТЗ-14

+++ src/main/java/ru/practicum/shareit/booking/BookingController.java
@@ -19,35 +19,35 @@ public class BookingController {
 
     @PostMapping
     public BookingDto create(@RequestBody BookingDto bookingDto,
-                             @RequestHeader(USER_ID) Integer bookerId) {
+                             @RequestHeader(USER_ID) Long bookerId) {
         return bookingService.addNewBooking(bookingDto, bookerId);
     }
 
     @PatchMapping("/{bookingId}")
-    public BookingDto update(@PathVariable Integer bookingId,
-                             @RequestHeader(USER_ID) Integer userId, @RequestParam BookingStatus status) {
+    public BookingDto update(@PathVariable Long bookingId,
+                             @RequestHeader(USER_ID) Long userId, @RequestParam BookingStatus status) {
         return bookingService.update(bookingId, userId, status);
     }
 
     @GetMapping("/{bookingId}")
-    public BookingDto getBookingById(@PathVariable Integer bookingId,
-                                     @RequestHeader(USER_ID) Integer userId) {
+    public BookingDto getBookingById(@PathVariable Long bookingId,
+                                     @RequestHeader(USER_ID) Long userId) {
         return bookingService.getBookingById(bookingId, userId);
     }
 
     @GetMapping
-    public List<BookingDto> getBookings(@RequestHeader(USER_ID) Integer userId) {
+    public List<BookingDto> getBookings(@RequestHeader(USER_ID) Long userId) {
         return bookingService.getAllBookings(userId);
     }
 
     @GetMapping("/{bookerId}")
-    public List<BookingDto> getBookingsByBooker(@RequestHeader(USER_ID) Integer bookerId) {
+    public List<BookingDto> getBookingsByBooker(@RequestHeader(USER_ID) Long bookerId) {
         return bookingService.getBookingsByBooker(bookerId);
     }
 
     @GetMapping("/{itemId}")
-    public List<BookingDto> getBookingByItem(@PathVariable Integer itemId,
-                                             @RequestHeader(USER_ID) Integer bookerId) {
+    public List<BookingDto> getBookingByItem(@PathVariable Long itemId,
+                                             @RequestHeader(USER_ID) Long bookerId) {
         return bookingService.getBookingByItem(itemId, bookerId);
     }
 }
