+++ src/main/java/ru/practicum/shareit/request/ItemRequestController.java
@@ -18,29 +18,29 @@ public class ItemRequestController {
 
     @PostMapping
     public ItemRequestDto create(@RequestBody ItemRequestDto itemRequestDto,
-                                 @RequestHeader(USER_ID) Integer requesterId) {
+                                 @RequestHeader(USER_ID) Long requesterId) {
         return itemRequestService.addRequest(itemRequestDto, requesterId);
     }
 
     @GetMapping("/{itemRequestId}")
-    public ItemRequestDto getItemRequestById(@PathVariable Integer itemRequestId,
-                                             @RequestHeader(USER_ID) Integer requesterId) {
+    public ItemRequestDto getItemRequestById(@PathVariable Long itemRequestId,
+                                             @RequestHeader(USER_ID) Long requesterId) {
         return itemRequestService.getRequestById(itemRequestId);
     }
 
     @GetMapping("/{requesterId}")
-    public List<ItemRequestDto> getItemRequestsByRequester(@RequestHeader(USER_ID) Integer requesterId) {
+    public List<ItemRequestDto> getItemRequestsByRequester(@RequestHeader(USER_ID) Long requesterId) {
         return itemRequestService.getItemsByRequester(requesterId);
     }
 
     @PatchMapping("/{itemRequestId}")
-    public ItemRequestDto update(@RequestBody ItemRequestDto itemRequestDto, @PathVariable Integer itemId,
-                                 @RequestHeader(USER_ID) Integer requesterId) {
+    public ItemRequestDto update(@RequestBody ItemRequestDto itemRequestDto, @PathVariable Long itemId,
+                                 @RequestHeader(USER_ID) Long requesterId) {
         return itemRequestService.update(itemRequestDto, itemId, requesterId);
     }
 
     @DeleteMapping("/{itemRequestId}")
-    public void delete(@PathVariable Integer itemId, @RequestHeader(USER_ID) Integer requesterId) {
+    public void delete(@PathVariable Long itemId, @RequestHeader(USER_ID) Long requesterId) {
         itemRequestService.delete(itemId, requesterId);
     }
 }
